requirements.txt
# Core
fastapi==0.115.2
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
python-dotenv==1.0.1

# AI/ML
openai==1.52.0

# HTTP Client
httpx==0.27.2

# Celery + Redis
celery==5.4.0
redis==5.0.1

# Email
# (using standard library imaplib for IMAP)
google-api-python-client==2.111.0
google-auth==2.25.2

# Document Processing
pypdf==4.0.1
openpyxl==3.1.2
python-docx==1.1.0
beautifulsoup4==4.12.2

# OCR
pytesseract==0.3.10
pdf2image==1.16.3
Pillow>=10.4.0

# Storage
boto3==1.34.14
azure-storage-blob==12.19.0

# WebSocket
websockets==12.0

# Security
# (ClamAV uses socket, no special package needed)

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
isort==5.13.2
mypy==1.8.0
